# orchestrator/server/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        description: Error 人类可读的错误说明
        example: insufficient permissions
        type: string
      success:
        description: Success 固定为 false
        example: false
        type: boolean
    type: object
...
host: localhost:9527
info:
  contact: {}
  description: |-
    Wingman 远程中控编排器 HTTP API（auth / RBAC / agent / script / workflow / messages / settings）。
    架构约束：Dashboard 与远程客户端只连接本 server，不直连 runtime。
  title: Wingman Orchestrator API
  version: "1.0"
paths:
  /admin/permissions:
    get:
      description: 返回全部权限码目录，支持按分类过滤；需要 users:manage 或 roles:manage 权限
      parameters:
      - description: 按分类过滤
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 权限目录（resource:action）
      tags:
      - admin-roles
  /admin/roles:
    get:
      description: 返回全部角色（内置优先）；需要 roles:manage 权限（admin 自动放行）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 角色列表
      tags:
      - admin-roles
    post:
      consumes:
      - application/json
      description: 新建自定义角色并绑定权限（内置 admin code 保留）；需要 roles:manage 权限
      parameters:
      - description: 角色定义
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - admin-roles
  /admin/roles/{code}:
    delete:
      description: 删除自定义角色并清理权限关联；内置角色或仍被用户引用时拒绝；需要 roles:manage 权限
      parameters:
      - description: 角色 code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - admin-roles
    get:
      description: 按 code（或兼容 :id 形式）返回单个角色及其权限
      parameters:
      - description: 角色 code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 角色详情
      tags:
      - admin-roles
    put:
      consumes:
      - application/json
      description: 自定义角色可改名称/描述，权限可更新（admin 角色保持通配不变）；需要 roles:manage 权限
      parameters:
      - description: 角色 code
        in: path
        name: code
        required: true
        type: string
      - description: 角色更新
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - admin-roles
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <JWT>"（由 POST /api/v1/auth/login 获取）'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"